+++ aws-cdn-templates/dev/env-cdn.sh
@@ -1,6 +1,0 @@
-export ZONE_ID="Z00641601ZB5KOCPK6AEX"
-export PORTALE_PA_CERTIFICATE_ARN="arn:aws:acm:us-east-1:558518206506:certificate/dc1d20b5-9f2f-4fad-ba53-3c1240b1b829"
-export PORTALE_PF_CERTIFICATE_ARN="arn:aws:acm:us-east-1:558518206506:certificate/5eaffa60-c73f-4844-b93f-153f94c3ab98"
-export PORTALE_PF_LOGIN_CERTIFICATE_ARN="arn:aws:acm:us-east-1:558518206506:certificate/d76e672d-77af-4dfc-bc62-0f1290ca3600"
-export PORTALE_PG_CERTIFICATE_ARN="arn:aws:acm:us-east-1:558518206506:certificate/3c7d089a-a052-4492-8089-1dd58c5f41c7"
-export REACT_APP_URL_API="https://webapi.dev.pn.pagopa.it/ https://pn-safestorage-eu-south-1-946373734005.s3.eu-south-1.amazonaws.com/"