+++ packages/pn-pa-webapp/src/navigation/__test__/navigation.utility.test.ts
@@ -1,7 +1,8 @@
 import { vi } from 'vitest';
 
 import { getConfiguration } from '../../services/configuration.service';
-import { goToSelfcareLogin } from '../navigation.utility';
+import { goToSelfcareLogin, goToSelfcareLogout } from '../navigation.utility';
+import { SELFCARE_LOGIN_PATH, SELFCARE_LOGOUT_PATH } from '../routes.const';
 
 const mockOpenFn = vi.fn();
 
@@ -15,6 +16,10 @@ describe('Tests notification.utility', () => {
     });
   });
 
+  beforeEach(() => {
+    mockOpenFn.mockClear();
+  });
+
   afterAll((): void => {
     Object.defineProperty(window, 'open', { configurable: true, value: original });
   });
@@ -22,6 +27,14 @@ describe('Tests notification.utility', () => {
   it('goToSelfcareLogin', () => {
     goToSelfcareLogin();
     expect(mockOpenFn).toHaveBeenCalledTimes(1);
-    expect(mockOpenFn).toHaveBeenCalledWith(getConfiguration().SELFCARE_URL_FE_LOGIN, '_self');
+    const url = `${getConfiguration().SELFCARE_BASE_URL}${SELFCARE_LOGIN_PATH}`
+    expect(mockOpenFn).toHaveBeenCalledWith(url, '_self');
+  });
+  
+  it('goToSelfcareLogout', () => {
+    goToSelfcareLogout();
+    expect(mockOpenFn).toHaveBeenCalledTimes(1);
+    const url = `${getConfiguration().SELFCARE_BASE_URL}${SELFCARE_LOGOUT_PATH}`
+    expect(mockOpenFn).toHaveBeenCalledWith(url, '_self');
   });
 });