+++ packages/pn-pa-webapp/src/navigation/navigation.utility.ts
@@ -1,7 +1,12 @@
 import { getConfiguration } from '../services/configuration.service';
+import { SELFCARE_LOGIN_PATH, SELFCARE_LOGOUT_PATH } from './routes.const';
 
 export function goToSelfcareLogin(): void {
-  /* eslint-disable functional/immutable-data */
-  const { SELFCARE_URL_FE_LOGIN } = getConfiguration();
-  window.open(SELFCARE_URL_FE_LOGIN, '_self');
+  const { SELFCARE_BASE_URL } = getConfiguration();
+  window.open( `${SELFCARE_BASE_URL}${SELFCARE_LOGIN_PATH}`, '_self');
+}
+
+export function goToSelfcareLogout(): void {
+  const { SELFCARE_BASE_URL } = getConfiguration();
+  window.open(`${SELFCARE_BASE_URL}${SELFCARE_LOGOUT_PATH}`, '_self');
 }