+++ packages/pn-personafisica-webapp/src/components/Contacts/LegalContacts.tsx
@@ -91,7 +91,8 @@ const LegalContacts = () => {
   const hasPecActive = defaultPECAddress?.value && defaultPECAddress.pecValid === true;
   const hasSercqSendActive = !!defaultSERCQ_SENDAddress;
   const isActive = hasPecActive || (hasSercqSendActive && !isValidatingPec);
-  const showSpecialContactsSection = specialAddresses.length > 0;
+  const showSpecialContactsSection =
+    specialAddresses.filter((addr) => addr.addressType === AddressType.LEGAL).length > 0;
 
   const channelType =
     hasSercqSendActive && !isValidatingPec ? ChannelType.SERCQ_SEND : ChannelType.PEC;
@@ -217,7 +218,7 @@ const LegalContacts = () => {
           {t(`legal-contacts.pec-disclaimer`, { ns: 'recapiti' })}
         </Alert>
       )}
-      {showSpecialContactsSection && <SpecialContacts />}
+      {showSpecialContactsSection && <SpecialContacts addressType={AddressType.LEGAL} />}
       <DeleteDialog
         showModal={modalOpen}
         removeModalTitle={t(