+++ packages/pn-personagiuridica-webapp/src/api/auth/Auth.api.ts
@@ -1,6 +1,6 @@
 import { TokenExchangeBody, User } from '../../models/User';
 import { authClient } from '../apiClients';
-import { AUTH_TOKEN_EXCHANGE } from './auth.routes';
+import { AUTH_LOGOUT, AUTH_TOKEN_EXCHANGE } from './auth.routes';
 
 export const AuthApi = {
   exchangeToken: async (spidToken: string, aar?: string): Promise<User> => {
@@ -41,4 +41,6 @@ export const AuthApi = {
     }
     return user;
   },
+  logout: (token: string): Promise<void> =>
+    authClient.post(AUTH_LOGOUT(), null, { headers: { Authorization: `Bearer ${token}` } }),
 };