+++ packages/pn-personagiuridica-webapp/src/utility/AppError/__test__/PGAppErrorFactory.test.ts
@@ -1,13 +1,18 @@
-import { AppError, ServerResponseError } from '@pagopa-pn/pn-commons';
+import { AppError, ServerResponseError, UnknownAppError } from '@pagopa-pn/pn-commons';
 
-import { PGAppErrorFactory } from '../PGAppErrorFactory';
-import { ServerResponseErrorCode } from '../types';
+import { DeliveryFileInfoNotFoundAppError } from '../DeliveryFileInfoNotFoundAppError';
+import { DeliveryNotificationWithoutPaymentAttachmentAppError } from '../DeliveryNotificationWithoutPaymentAttachmentAppError';
+import { DeliveryPushFileNotFoundAppError } from '../DeliveryPushFileNotFoundAppError';
+import { GenericInvalidParameterPatternAppError } from '../GenericInvalidParameterPatternAppError';
+import { InvalidBodyAppError } from '../InvalidBodyAppError';
 import { MandateAlreadyExistsAppError } from '../MandateAlreadyExistsAppError';
 import { MandateDelegateHimselfAppError } from '../MandateDelegateHimselfAppError';
 import { MandateInvalidVerificationCodeAppError } from '../MandateInvalidVerificationCodeAppError';
 import { MandateNotAcceptableAppError } from '../MandateNotAcceptableAppError';
 import { MandateNotFoundAppError } from '../MandateNotFoundAppError';
+import { PGAppErrorFactory } from '../PGAppErrorFactory';
 import { UserAttributesInvalidVerificationCodeAppError } from '../UserAttributesInvalidVerificationCodeAppError';
+import { ServerResponseErrorCode } from '../types';
 
 class PGAppErrorFactoryForTest extends PGAppErrorFactory {
   constructor(translateFunction: (path: string, ns: string) => string) {
@@ -63,5 +68,45 @@ describe('Test PGAppErrorFactory', () => {
       code: ServerResponseErrorCode.PN_USERATTRIBUTES_INVALIDVERIFICATIONCODE,
     });
     expect(errorClass).toBeInstanceOf(UserAttributesInvalidVerificationCodeAppError);
+  });
+
+  it('Should return GenericInvalidParameterPatternAppError', () => {
+    const errorClass = errorFactory.getCustomErrorForTest({
+      code: ServerResponseErrorCode.PN_GENERIC_INVALIDPARAMETER_PATTERN,
+    });
+    expect(errorClass).toBeInstanceOf(GenericInvalidParameterPatternAppError);
+  });
+
+  it('Should return InvalidBodyAppError', () => {
+    const errorClass = errorFactory.getCustomErrorForTest({
+      code: ServerResponseErrorCode.PN_INVALID_BODY,
+    });
+    expect(errorClass).toBeInstanceOf(InvalidBodyAppError);
+  });
+
+  it('Should return DeliveryFileInfoNotFoundAppError', () => {
+    const errorClass = errorFactory.getCustomErrorForTest({
+      code: ServerResponseErrorCode.PN_DELIVERY_FILEINFONOTFOUND,
+    });
+    expect(errorClass).toBeInstanceOf(DeliveryFileInfoNotFoundAppError);
+  });
+
+  it('Should return DeliveryNotificationWithoutPaymentAttachmentAppError', () => {
+    const errorClass = errorFactory.getCustomErrorForTest({
+      code: ServerResponseErrorCode.PN_DELIVERY_NOTIFICATIONWITHOUTPAYMENTATTACHMENT,
+    });
+    expect(errorClass).toBeInstanceOf(DeliveryNotificationWithoutPaymentAttachmentAppError);
+  });
+
+  it('Should return DeliveryPushFileNotFoundAppError', () => {
+    const errorClass = errorFactory.getCustomErrorForTest({
+      code: ServerResponseErrorCode.PN_DELIVERYPUSH_FILE_NOT_FOUND,
+    });
+    expect(errorClass).toBeInstanceOf(DeliveryPushFileNotFoundAppError);
+  });
+
+  it('Should return UnknownAppError', () => {
+    const errorClass = errorFactory.getCustomErrorForTest({ code: 'UNKNOWN_CODE' });
+    expect(errorClass).toBeInstanceOf(UnknownAppError);
   });
 });